Make sure app.cached_users is initialized
Before this, because apache starts up the code at the same time as the
first request, it was possible for the cron scheduler to not complete
its run of update_groups() before the POST handler runs, making it
think the request is not allowed even if should be.

This would then cause a stacktrace in the jenkins sua package build when
running sendSaltEvent()

With this change we synchronously fill in the dict before the request
handler can run for the first time.

Debugged and discussed with @simmel.